refactor(blorktools): threejs scene util
